Add files via upload
I tried three other times to make this project as a full-blown library that was opaque, used iterative macros, and tons of functions for working with the kind of intrusive doubly linked lists used in the Linux kernel. I learned a lot with every attempt, but each time I ended up bricking the library and deciding to start over. Only after getting rid of the macros, going from opaque to open-source, and putting everything in a single monolithic file did it work more or less. Intrusive is a kind of linked list where you embed the "library" node inside of another node presumably a user-defined, application-specific data structure. If you want to change the list association of your struct, you just change the type of embedded node used (or use an additional embedded node). If you want to access the data from the member pointer, use container_of() and offsetof() macros to return a pointer to the parent container. Kind of neat and I liked the self-contained feel of it. I learned that iterative macros are really cool but very difficult. I also learned that if I want to use opaque pointer design, I'll probably just switch to C++ and use the keyword "private", rather than going crazy with makefiles, macro-wrapped functions, and making complicated arrangements of nested internal header & source files...